Updated main program (11).
Integrated the confirm_cancel() function into the main program. Made it so that when the order is confirmed for click and collect, the print statement of a text message to be received when their order is ready for click and collect will be printed. If the order is confirmed for delivery, the print statement of a text message to be received of the delivery status of the order will be printed. The cancel option for the order works properly.