Updated copy of main program (5).
Added a while loop using the time module so that at the very beginning of the program, a loading message will update itself in the same line it appears in, having a dot appear after the 'Loading Program' message, each second until three dots appear. After this, the rest of the program will continue.