Updated copy of main program (13), edited/formatted text.
Edited a question in the choosing albums to purchase section of the program. Added another while loop for an updating print statement like the loading program animation at the start of the program but for printing the order out. Edited some texts in the print order section of the program.